% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_vowels.R
\name{count_vowels}
\alias{count_vowels}
\title{Policz samogłoski a/e/i/o/u/y.}
\usage{
count_vowels(str)
}
\arguments{
\item{str}{Ciąg wejściowy.}
}
\value{
Liczba samogłosek.
}
\description{
Policz samogłoski a/e/i/o/u/y.
}
