% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp-wrapper.R
\name{word_len_cpp}
\alias{word_len_cpp}
\title{Długości słów (C++ via Rcpp).}
\usage{
word_len_cpp(strings)
}
\arguments{
\item{strings}{Wektor znaków.}
}
\value{
Wektor długości.
}
\description{
Długości słów (C++ via Rcpp).
}
